Dedupe static vs draggable.
